--- public/DataSetSample.tsv
@@ -1,2 +1,0 @@
-Dataset Name	Data Type	Species	Phenotype/Indication	# of participants	Description	dbGAP	Data Depositor	Principal Investigator(PI)	Sample Collection ID	Consent ID
-(Bucienne Monco) - Muc-1 Kidney Disease	DNA, whole genome	human	muc-1, kidney disease	31	muc-1 patients that developed cancer , 5 weeks after treatment	http://....	John Doe	Mark Smith	SC-20658
