--- public/config-example.json
@@ -1,16 +1,0 @@
-{
-  "env": "local|dev|staging|prod",
-  "tag": "dev",
-  "hash": "dev",
-  "apiUrl": "http://localhost:8180/api",
-  "ontologyApiUrl": "https://ontologyURL.org/",
-  "clientId": "111111111111-11111111111111111111111111111111.apps.googleusercontent.com",
-  "errorApiKey": "example",
-  "gaId": "",
-  "profileUrl": "https://profile-dot-broad-shibboleth-prod.appspot.com/dev",
-  "nihUrl": "https://broad-shibboleth-prod.appspot.com/dev/login",
-  "features": {
-    "featureX" : true,
-    "featureY" : false
-  }
-}
