--- public/images/favicon/browserconfig.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<browserconfig>
-    <msapplication>
-        <tile>
-            <square150x150logo src="%PUBLIC_URL%/images/favicon/mstile-150x150.png"/>
-            <TileColor>#2b5797</TileColor>
-        </tile>
-    </msapplication>
-</browserconfig>
